# contracts/jarvis-agent.openapi.json
{
  "components": {
    "schemas": {
      "ActionDecision": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "enum": [
              "approve",
              "deny"
            ],
            "title": "Decision",
            "type": "string"
          },
          "payload_hash": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Payload Hash",
            "type": "string"
          },
          "session_id": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "payload_hash",
          "decision"
        ],
        "title": "ActionDecision",
        "type": "object"
      },
      "ActionEnvelope": {
        "additionalProperties": false,
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ActionResponse"
          }
        },
        "required": [
          "result"
        ],
        "title": "ActionEnvelope",
        "type": "object"
      },
...
      "AgentEventPollResponse": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/AgentEventResponse"
            },
            "title": "Events",
            "type": "array"
          },
          "next_after": {
            "title": "Next After",
            "type": "integer"
          }
        },
        "required": [
          "events",
          "next_after"
        ],
        "title": "AgentEventPollResponse",
        "type": "object"
      },
...
      "ContextDelete": {
        "additionalProperties": false,
        "properties": {
          "codex_thread_id": {
            "default": "",
            "maxLength": 256,
            "title": "Codex Thread Id",
            "type": "string"
          },
          "project_key": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Project Key",
            "type": "string"
          }
        },
        "required": [
          "project_key"
        ],
        "title": "ContextDelete",
        "type": "object"
      },
      "ContextDeleteResponse": {
        "additionalProperties": false,
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "ContextDeleteResponse",
        "type": "object"
      },
      "ContextEnvelope": {
        "additionalProperties": false,
        "properties": {
          "context": {
            "$ref": "#/components/schemas/ContextResponse"
          }
        },
        "required": [
          "context"
        ],
        "title": "ContextEnvelope",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobEnvelope": {
        "additionalProperties": false,
        "properties": {
          "result": {
            "$ref": "#/components/schemas/JobResponse"
          }
        },
        "required": [
          "result"
        ],
        "title": "JobEnvelope",
        "type": "object"
      },
...
      "ManifestEntryResponse": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          }
        },
        "required": [
          "name",
          "description",
          "parameters"
        ],
        "title": "ManifestEntryResponse",
        "type": "object"
      },
      "ProposalCreate": {
        "additionalProperties": false,
        "properties": {
          "arguments": {
            "additionalProperties": true,
            "title": "Arguments",
            "type": "object"
          },
          "function_call_id": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Function Call Id",
            "type": "string"
          },
          "generation": {
            "exclusiveMinimum": 0.0,
            "title": "Generation",
            "type": "integer"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "turn_id": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turn Id"
          }
        },
        "required": [
          "generation",
          "function_call_id",
          "name"
        ],
        "title": "ProposalCreate",
        "type": "object"
      },
...
      "ProviderWebhookResponse": {
        "additionalProperties": false,
        "properties": {
          "disposition": {
            "title": "Disposition",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "status": {
            "const": "accepted",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "disposition",
          "event_id"
        ],
        "title": "ProviderWebhookResponse",
        "type": "object"
      },
      "SessionClose": {
        "additionalProperties": false,
        "properties": {
          "generation": {
            "exclusiveMinimum": 0.0,
            "title": "Generation",
            "type": "integer"
          }
        },
        "required": [
          "generation"
        ],
        "title": "SessionClose",
        "type": "object"
      },
      "SessionCloseResponse": {
        "additionalProperties": false,
        "properties": {
          "generation": {
            "title": "Generation",
            "type": "integer"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "state": {
            "const": "CLOSED",
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "generation",
          "state"
        ],
        "title": "SessionCloseResponse",
        "type": "object"
      },
      "SessionCreate": {
        "additionalProperties": false,
        "properties": {
          "codex_thread_id": {
            "default": "",
            "maxLength": 256,
            "title": "Codex Thread Id",
            "type": "string"
          },
          "project_key": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Project Key",
            "type": "string"
          }
        },
        "required": [
          "project_key"
        ],
        "title": "SessionCreate",
        "type": "object"
      },
      "SessionResponse": {
        "additionalProperties": false,
        "properties": {
          "catalog": {
            "$ref": "#/components/schemas/CatalogResponse"
          },
          "context": {
            "$ref": "#/components/schemas/ContextResponse"
          },
          "generation": {
            "title": "Generation",
            "type": "integer"
          },
          "manifest": {
            "items": {
              "$ref": "#/components/schemas/ManifestEntryResponse"
            },
            "title": "Manifest",
            "type": "array"
          },
          "manifest_version": {
            "title": "Manifest Version",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "state": {
            "enum": [
              "ACTIVE",
              "CLOSED"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "generation",
          "state",
          "manifest_version",
          "manifest",
          "catalog",
          "context"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "TurnCreate": {
        "additionalProperties": false,
        "properties": {
          "final": {
            "const": true,
            "title": "Final",
            "type": "boolean"
          },
          "generation": {
            "exclusiveMinimum": 0.0,
            "title": "Generation",
            "type": "integer"
          },
          "transcript": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Transcript",
            "type": "string"
          },
          "turn_id": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Turn Id",
            "type": "string"
          }
        },
        "required": [
          "generation",
          "turn_id",
          "transcript",
          "final"
        ],
        "title": "TurnCreate",
        "type": "object"
      },
      "TurnResponse": {
        "additionalProperties": false,
        "properties": {
          "project_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Key"
          },
          "status": {
            "enum": [
              "committed",
              "duplicate"
            ],
            "title": "Status",
            "type": "string"
          },
          "transcript_hash": {
            "title": "Transcript Hash",
            "type": "string"
          },
          "turn_id": {
            "title": "Turn Id",
            "type": "string"
          }
        },
        "required": [
          "turn_id",
          "status",
          "transcript_hash"
        ],
        "title": "TurnResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenJarvis Jarvis Agent API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/jarvis/agent/actions/{action_id}": {
      "get": {
        "operationId": "get_action_v1_jarvis_agent_actions__action_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Action",
        "tags": [
          "jarvis-agent"
        ]
      }
    },
    "/v1/jarvis/agent/actions/{action_id}/decision": {
      "post": {
        "operationId": "decide_action_v1_jarvis_agent_actions__action_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "title": "Action Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Jarvis-Decision-Channel",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Jarvis-Decision-Channel",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Action",
        "tags": [
          "jarvis-agent"
        ]
      }
    },
...
    "/v1/jarvis/agent/context": {
      "delete": {
        "operationId": "delete_context_v1_jarvis_agent_context_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextDelete"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Context",
        "tags": [
          "jarvis-agent"
        ]
      },
      "get": {
        "operationId": "get_context_v1_jarvis_agent_context_get",
        "parameters": [
          {
            "in": "query",
            "name": "project_key",
            "required": true,
            "schema": {
              "maxLength": 1024,
              "minLength": 1,
              "title": "Project Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "codex_thread_id",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 256,
              "title": "Codex Thread Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Context",
        "tags": [
          "jarvis-agent"
        ]
      }
    },
...
    "/v1/jarvis/agent/events": {
      "get": {
        "operationId": "events_v1_jarvis_agent_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Events",
        "tags": [
          "jarvis-agent"
        ]
      }
    },
    "/v1/jarvis/agent/events/poll": {
      "get": {
        "description": "Return a finite event page for transports that cannot carry SSE.",
        "operationId": "poll_events_v1_jarvis_agent_events_poll_get",
        "parameters": [
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentEventPollResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Poll Events",
        "tags": [
          "jarvis-agent"
        ]
      }
    },
    "/v1/jarvis/agent/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_v1_jarvis_agent_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jarvis-agent"
        ]
      }
    },
    "/v1/jarvis/agent/providers/acelerachat/webhooks": {
      "post": {
        "operationId": "acelerachat_webhook_v1_jarvis_agent_providers_acelerachat_webhooks_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-AceleraChat-Delivery",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Acelerachat-Delivery",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-AceleraChat-Timestamp",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Acelerachat-Timestamp",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-AceleraChat-Signature",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Acelerachat-Signature",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-AceleraChat-Signature-Previous",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Acelerachat-Signature-Previous",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderWebhookResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Acelerachat Webhook",
        "tags": [
          "jarvis-agent"
        ]
      }
    },
    "/v1/jarvis/agent/sessions": {
      "post": {
        "operationId": "create_session_v1_jarvis_agent_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "jarvis-agent"
        ]
      }
    },
    "/v1/jarvis/agent/sessions/{session_id}/close": {
      "post": {
        "operationId": "close_session_v1_jarvis_agent_sessions__session_id__close_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionClose"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCloseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close Session",
        "tags": [
          "jarvis-agent"
        ]
      }
    },
    "/v1/jarvis/agent/sessions/{session_id}/proposals": {
      "post": {
        "operationId": "create_proposal_v1_jarvis_agent_sessions__session_id__proposals_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Proposal",
        "tags": [
          "jarvis-agent"
        ]
      }
    },
    "/v1/jarvis/agent/sessions/{session_id}/turns": {
      "post": {
        "operationId": "commit_turn_v1_jarvis_agent_sessions__session_id__turns_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TurnCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commit Turn",
        "tags": [
          "jarvis-agent"
        ]
      }
    }
  }
}